+++ src/imio/smartweb/core/webcomponents/src/components/Events/ContactContent/ContactContent.jsx
@@ -4,21 +4,25 @@ import useAxios from "../../../hooks/useAxios";
 import useFilterQuery from "../../../hooks/useFilterQuery";
 import moment from 'moment'
 import Moment from 'react-moment';
-const ContactContent = ({ queryUrl, onChange }) => {
+const ContactContent = ({queryUrl, onChange }) => {
     let history = useHistory();
     const queryString = require("query-string");
     const parsed = queryString.parse(useFilterQuery().toString());
-    const [params, setParams] = useState({});
+    const parsed2 = { ...parsed, UID: parsed['u'], fullobjects: 1 };
+    const [params, setParams] = useState(parsed2);
     const [contactItem, setcontactItem] = useState({});
-    const { response, error, isLoading } = useAxios({
-        method: "get",
-        url: "",
-        baseURL: queryUrl,
-        headers: {
-            Accept: "application/json",
+    const { response, error, isLoading } = useAxios(
+        {
+            method: "get",
+            url: "",
+            baseURL: queryUrl,
+            headers: {
+                Accept: "application/json",
+            },
+            params: params,
         },
-        params: params,
-    });
+        []
+    );
 
     // set all contacts state
     useEffect(() => {
@@ -27,19 +31,12 @@ const ContactContent = ({ queryUrl, onChange }) => {
         }
     }, [response]);
 
-    useEffect(() => {
-        setParams({
-            UID: parsed.u,
-            fullobjects: 1,
-        });
-    }, []);
-
-
     function handleClick() {
-        history.push(".");
+        history.push("./");
         onChange(null);
     }
-
+    const start = moment(contactItem.start);
+    const end = moment(contactItem.end);
     return (
         <div className="envent-content r-content">
             <button type="button" onClick={handleClick}>
@@ -59,17 +56,27 @@ const ContactContent = ({ queryUrl, onChange }) => {
                         }}
                     />
                 </figure>
-                {/* <div class="r-content-date">
-                    <div className="r-content-date-publish">
-                        <span>Publié le </span>
-                        <Moment format='DD-MM-YYYY'>{publish}</Moment>
+                <div className="r-content-news-info"> 
+                    <div class="r-content-date">
+                        <div className="r-content-date-start">
+                            <span>Du </span>
+                            <Moment format='DD-MM-YYYY'>{start}</Moment>
+                        </div>
+                        <div className="r-content-date-end">
+                            <span>au </span>
+                            <Moment format='DD-MM-YYYY'>{end}</Moment>
+                        </div>
                     </div>
-                    <div className="r-content-date-last">
-                        <span>Modifié le</span>
-                        <Moment format='DD-MM-YYYY'>{lastModified}</Moment>
+                    <div className="r-content-book">
+                        <a href={contactItem.ticket_url}>Billetterie</a>
                     </div>
-                </div> */}
-
+                    <div className="r-content-event-link">
+                        <a href={contactItem.event_url}>Lien de l'événement</a>
+                    </div>
+                    <div className="r-content-online-participation">
+                        <a href={contactItem.online_participation}>Participation en ligne</a>
+                    </div>
+                </div>
                 <div class="r-content-text"
                     dangerouslySetInnerHTML={{
                         __html: contactItem.text && contactItem.text.data