+++ src/imio/smartweb/core/contents/sections/html/content.py
@@ -36,3 +36,5 @@ class ISectionHTML(ISection):
 @implementer(ISectionHTML)
 class SectionHTML(Section):
     """SectionHTML class"""
+
+    manage_content = True